Grund SQL und Room Klasse erstellt, LandingPageController und BuildingRepository dementsprechend angepasst.

+++ src/main/java/com/example/FinkenauNavigator/BuildingRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-interface BuildingRepository extends ListCrudRepository<Building, Long> {
+interface BuildingRepository extends ListCrudRepository<Building, Integer> {
 
     @Query("SELECT * FROM BUILDING WHERE LOWER(name) = LOWER(:name)")
     List<Building> findRoomByName(String name);
